Refactored Gz/Gazebo launch to dynamically spawn bcr_bot
